# http/api/go/model_created_cat.go
/*
 * Pets API
 *
 * A simple API to manage pets
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api




type CreatedCat struct {

	Id string `json:"id,omitempty"`

	Name string `json:"name"`

	Breed string `json:"breed"`

	Age int32 `json:"age"`
}

// AssertCreatedCatRequired checks if the required fields are not zero-ed
func AssertCreatedCatRequired(obj CreatedCat) error {
	elements := map[string]interface{}{
		"name": obj.Name,
		"breed": obj.Breed,
		"age": obj.Age,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertCreatedCatConstraints checks if the values respects the defined constraints
func AssertCreatedCatConstraints(obj CreatedCat) error {
	return nil
}