# http/api/go/model_updated_cat.go
/*
 * Pets API
 *
 * A simple API to manage pets
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api




type UpdatedCat struct {

	Id string `json:"id,omitempty"`

	Name string `json:"name"`

	Breed string `json:"breed"`

	Age int32 `json:"age"`
}

// AssertUpdatedCatRequired checks if the required fields are not zero-ed
func AssertUpdatedCatRequired(obj UpdatedCat) error {
	elements := map[string]interface{}{
		"name": obj.Name,
		"breed": obj.Breed,
		"age": obj.Age,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertUpdatedCatConstraints checks if the values respects the defined constraints
func AssertUpdatedCatConstraints(obj UpdatedCat) error {
	return nil
}